[chore] hardware customScheme 으로 Entry-hw 실행 부분 로직 추출
- 이름이 ieLauncher 인데 모든 브라우저 커버하고 있어서 이름 수정
  - 이름 수정하면서 그에 맞춰 코드 수정
- 헝가리안 표기 변수명을 카멜 표기법으로 수정
